fix(import): name the files that failed, and say which failure it was
Grouping by reason solved the 500-file case and broke the small one: any
group of two or more named nothing, so a five-file import with two failures
said "2 files — could not be read" and the log line that would identify them
is redacted. Names are listed while the group is small.

That interacted badly with every non-`ImportError` collapsing into one
bucket. A permissions failure, a false-positive binary rejection and an I/O
error rendered as one line with no filenames anywhere the user could read
them. The common Cocoa read errors are now distinguished.

The binary heuristic had no absolute floor, so a bare ratio meant a single
control character rejected any file of twenty scalars or fewer — a one-line
note pasted from a terminal was refused outright, and whether a pasted build
log imported depended on how much prose happened to surround it. It now
requires both a floor and the ratio.

A partial read says so, and a cancelled import says that instead of
reporting nothing imported as though the folder had been empty.

`SpaceTreeRow` moves to its own file: the reporting pushed
`CategorySidebarView.swift` past the length limit, and it was two types in
one file already.
